+++ R/paths.R
@@ -1,4 +1,3 @@
-
 corePaths <- c("modulePath", "cachePath", "inputPath", "outputPath")
 tmpPaths <- c("rasterPath", "scratchPath", "terraPath")
 spPaths <- c(corePaths, tmpPaths)